fix: 修复了一些Review过程中出现的问题
相关的修改:
1.部分函数名字存在低级的拼写错误
2.无用的command命令
3.无效赋值
4.过多的using
5.修复了删除用户时可能会出现意外的删除

+++ SpringKey/Files/LinkFile.cs
@@ -56,8 +56,16 @@ namespace SpringKey.Files
         {
             if (File.Exists(FilePath))
                 File.Delete(FilePath);
-            if (Directory.Exists(UserPath))
+            if (Directory.Exists(UserPath) && IsPathSafe(UserPath))
                 Directory.Delete(UserPath, true);
+        }
+
+        private bool IsPathSafe(string path)
+        {
+            var full = Path.GetFullPath(path);
+            var basePath = Path.GetFullPath(DataPath);
+            return full.StartsWith(basePath + Path.DirectorySeparatorChar)
+                   || full == basePath;
         }
 
         public void Rename(string newUserName)
